Move transform fn to internal module
