+++ src/main.rs
@@ -32,7 +32,6 @@ mod windows;
 
 use std::env;
 use std::error::Error;
-use std::ffi::OsString;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time;
@@ -41,7 +40,7 @@ use atty::Stream;
 use regex::RegexBuilder;
 
 use exec::CommandTemplate;
-use internal::{error, pattern_has_uppercase_char, FdOptions};
+use internal::{error, pattern_has_uppercase_char, transform_args_with_exec, FdOptions};
 use lscolors::LsColors;
 use walk::FileType;
 
@@ -171,120 +170,4 @@ fn main() {
         Ok(re) => walk::scan(&dir_vec, Arc::new(re), Arc::new(config)),
         Err(err) => error(err.description()),
     }
-}
-
-/// Traverse args_os, looking for -exec and replacing it with --exec.
-///
-/// # Returns
-///
-/// * The args, with substitution if required
-fn transform_args_with_exec<I>(original: I) -> Vec<OsString>
-where
-    I: Iterator<Item = OsString>,
-{
-    let mut in_exec_opt = false;
-    let target = OsString::from("-exec");
-    let long_start = OsString::from("--exec");
-    let short_start = OsString::from("-x");
-    let exec_end = OsString::from(";");
-
-    original.fold(vec![], |mut args, curr| {
-        if in_exec_opt {
-            if curr == exec_end {
-                in_exec_opt = false;
-            }
-            args.push(curr);
-            return args;
-        }
-
-        if curr == target || curr == long_start || curr == short_start {
-            args.push(if curr == target {
-                OsString::from("--exec")
-            } else {
-                curr
-            });
-            in_exec_opt = true;
-        } else {
-            args.push(curr);
-        }
-        args
-    })
-}
-
-#[cfg(test)]
-fn oss(v: &str) -> OsString {
-    OsString::from(v)
-}
-
-/// Ensure that -exec gets transformed into --exec
-#[test]
-fn normal_exec_substitution() {
-    let original = vec![oss("fd"), oss("foo"), oss("-exec"), oss("cmd")];
-    let expected = vec![oss("fd"), oss("foo"), oss("--exec"), oss("cmd")];
-
-    let actual = transform_args_with_exec(original.into_iter());
-    assert_eq!(expected, actual);
-}
-
-/// Ensure that --exec is not touched
-#[test]
-fn passthru_of_original_exec() {
-    let original = vec![oss("fd"), oss("foo"), oss("--exec"), oss("cmd")];
-    let expected = vec![oss("fd"), oss("foo"), oss("--exec"), oss("cmd")];
-
-    let actual = transform_args_with_exec(original.into_iter());
-    assert_eq!(expected, actual);
-}
-
-#[test]
-fn temp_check_that_exec_context_observed() {
-    let original = vec![
-        oss("fd"),
-        oss("foo"),
-        oss("-exec"),
-        oss("cmd"),
-        oss("-exec"),
-        oss("ls"),
-        oss(";"),
-        oss("-exec"),
-        oss("rm"),
-        oss(";"),
-        oss("--exec"),
-        oss("find"),
-        oss("-exec"),
-        oss("rm"),
-        oss(";"),
-        oss("-x"),
-        oss("foo"),
-        oss("-exec"),
-        oss("something"),
-        oss(";"),
-        oss("-exec"),
-    ];
-    let expected = vec![
-        oss("fd"),
-        oss("foo"),
-        oss("--exec"),
-        oss("cmd"),
-        oss("-exec"),
-        oss("ls"),
-        oss(";"),
-        oss("--exec"),
-        oss("rm"),
-        oss(";"),
-        oss("--exec"),
-        oss("find"),
-        oss("-exec"),
-        oss("rm"),
-        oss(";"),
-        oss("-x"),
-        oss("foo"),
-        oss("-exec"),
-        oss("something"),
-        oss(";"),
-        oss("--exec"),
-    ];
-
-    let actual = transform_args_with_exec(original.into_iter());
-    assert_eq!(expected, actual);
 }
